+++ PGP.WebUI/ClientApp/src/features/UsersPage.tsx
@@ -4,6 +4,7 @@ import { Avatar, Button, Spinner, Table, toaster } from 'evergreen-ui';
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
 import { deleteUser, getUsers } from '../apiClient';
+import { useDeleteConfirmation } from '../components/DeleteConfirmationService';
 import * as enums from '../enums';
 import { IUser } from '../models';
 import { routes } from '../utils/routes';
@@ -11,7 +12,9 @@ import { routes } from '../utils/routes';
 const UsersPage: React.FC = () => {
     const [users, setUsers] = useState<IUser[]>([]);
     const [isLoading, setIsLoading] = useState<boolean>(true);
+
     const history = useHistory();
+    const confirm = useDeleteConfirmation();
 
     useEffect(() => {
         const init = async () => {
@@ -25,13 +28,16 @@ const UsersPage: React.FC = () => {
     }, []);
 
     const handleDelete = async (userId: number) => {
-        const response = await deleteUser(userId);
+        confirm().then(async () => {
+            const response = await deleteUser(userId);
 
-        if (response) {
-            setUsers(users.filter(u => u.id !== userId));
-            toaster.success('User was successfully deleted.');
-        }
+            if (response) {
+                setUsers(users.filter(u => u.id !== userId));
+                toaster.success('User was successfully deleted.');
+            }
+        });
     };
+
     return (
         <div className="container mt-5 mb-5">
             {isLoading ? (
